requirements.txt
mkdocs-exclude<2
mkdocs-git-authors-plugin<1
mkdocs-git-committers-plugin-2<3
mkdocs-git-revision-date-localized-plugin<2
mkdocs-glightbox<1
mkdocs-material[imaging] @ git+https://${GH_TOKEN}@github.com/squidfunk/mkdocs-material-insiders.git@master
mkdocs-minify-plugin<1
mkdocs-rss-plugin<2
mkdocs-table-reader-plugin<3
psycopg2-binary<3
pymdown-extensions<11
